Hidden Links Page
Jaime may consider deleting the links page and moving the content to other places on the website. In the mean time, since the content of the page is not populated, I have hidden the links page by removing it from the navlinks data folder. The page still exists on the website, but it is not accessible without the URL.